Replace title url field with seed urls field
Previously the seed URL is hidden under gather options. This makes it
easy to forget to update when changing the title URL. It also makes it
not obvious when a title has multiple seed URLs.

Since seed URLs are actually what we crawl based on it's the thing we
actually care about most so let it be the primary input field. We also
now leave the title URL blank on the form when it's the same as the
primary seed URL. This means that updating the seed URL will
update any matching title URL.

Suggested-By: Paul Koerbin