TitleEdit: Experiment with using history.go(-2) to go back after saving
A user was expressing some frustration with doing searching (sometimes involved the use of filters) and not being able to get back to the search results easily after, say, editing a record. Basically you have to back button through a number of screens to get back to the search results.